requirements.txt
asgiref==3.10.0
blinker==1.9.0
cachetools==6.2.1
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
contourpy==1.3.3
cycler==0.12.1
Django==5.2.7
filelock==3.20.0
Flask==3.1.2
fonttools==4.60.1
fsspec==2025.9.0
google-api-core==2.28.1
google-api-python-client==2.186.0
google-auth==2.41.1
google-auth-httplib2==0.2.1
google-auth-oauthlib==1.2.3
googleapis-common-protos==1.71.0
httplib2==0.31.0
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
kiwisolver==1.4.9
MarkupSafe==3.0.3
matplotlib==3.10.7
mpmath==1.3.0
networkx==3.5
numpy==2.2.6
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.3
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvtx-cu12==12.8.90
oauthlib==3.3.1
opencv-python==4.12.0.88
packaging==25.0
pillow==11.3.0
polars==1.34.0
polars-runtime-32==1.34.0
proto-plus==1.26.1
protobuf==6.33.0
psutil==7.1.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pyparsing==3.2.5
python-dateutil==2.9.0.post0
PyYAML==6.0.3
requests==2.32.5
requests-oauthlib==2.0.0
rsa==4.9.1
scipy==1.16.2
setuptools==80.9.0
six==1.17.0
sqlparse==0.5.3
sympy==1.14.0
torch==2.8.0
torchvision==0.23.0
triton==3.4.0
typing_extensions==4.15.0
ultralytics==8.3.214
ultralytics-thop==2.0.17
uritemplate==4.2.0
urllib3==2.5.0
Werkzeug==3.1.3
yt-dlp==2025.10.14
